Revert "Merge branch 'rustrewrite' of https://github.com/Oki1/OpenGLGameEngine into rustrewrite"
This reverts commit 490c5b930342e06045a19c1d197ec08cbba266a2, reversing
changes made to 0270bb7748c18e25e1d555cc9501f9c2cd6b0f01.

+++ src/model/loader/mod.rs
@@ -32,7 +32,7 @@ fn read_f32_array(f: &mut File, n: &usize) -> std::io::Result<Box<[f32]>> {
 }
 
 fn read_i32_array(f: &mut File, n: &usize) -> std::io::Result<Box<[i32]>> {
-    let mut indices_raw = vec![0u8; *n * std::mem::size_of::<i32>()];
+    let mut indices_raw = vec![0u8; *n * std::mem::size_of::<f32>()];
     let mut indices_vec = vec![0i32; *n];
     f.read_exact(&mut indices_raw)?;
     LittleEndian::read_i32_into(&indices_raw, &mut indices_vec);
